:boom: :recycle: move extraction PDF and Image into their respective namespace
