--- mindee/extraction/__init__.py
@@ -1,17 +1,0 @@
-from mindee.extraction.common.extracted_image import ExtractedImage
-from mindee.extraction.common.image_extractor import (
-    attach_image_as_new_file,
-    extract_multiple_images_from_source,
-)
-from mindee.extraction.multi_receipts_extractor import multi_receipts_extractor
-from mindee.extraction.pdf_extractor.extracted_pdf import ExtractedPdf
-from mindee.extraction.pdf_extractor.pdf_extractor import PdfExtractor
-
-__all__ = [
-    "ExtractedImage",
-    "ExtractedPdf",
-    "PdfExtractor",
-    "attach_image_as_new_file",
-    "extract_multiple_images_from_source",
-    "multi_receipts_extractor",
-]
