--- mindee/extraction/multi_receipts_extractor/__init__.py
@@ -1,3 +1,0 @@
-from mindee.extraction.multi_receipts_extractor import multi_receipts_extractor
-
-__all__ = ["multi_receipts_extractor"]
