--- mindee/extraction/pdf_extractor/__init__.py
@@ -1,4 +1,0 @@
-from mindee.extraction.pdf_extractor.extracted_pdf import ExtractedPdf
-from mindee.extraction.pdf_extractor.pdf_extractor import PdfExtractor
-
-__all__ = ["ExtractedPdf", "PdfExtractor"]
