+++ mindee/image/extracted_image.py
