+++ mindee/v1/pdf/multi_receipts_extractor.py
@@ -1,10 +1,8 @@
 from typing import List
 
 from mindee.error.mindee_error import MindeeError
-from mindee.extraction.common.extracted_image import ExtractedImage
-from mindee.extraction.common.image_extractor import (
-    extract_multiple_images_from_source,
-)
+from mindee.image.extracted_image import ExtractedImage
+from mindee.image.image_extractor import extract_multiple_images_from_source
 from mindee.input.local_input_source import LocalInputSource
 from mindee.v1.parsing.common.inference import Inference
