+++ mindee/v2/file_operations/crop.py
@@ -1,8 +1,9 @@
 from typing import List, Union
 
 from mindee.error import MindeeError
-from mindee.extraction import ExtractedImage, extract_multiple_images_from_source
 from mindee.geometry import Point, Polygon
+from mindee.image.extracted_image import ExtractedImage
+from mindee.image.image_extractor import extract_multiple_images_from_source
 from mindee.input.local_input_source import LocalInputSource
 from mindee.v2.parsing.inference.field import FieldLocation
 from mindee.v2.file_operations.crop_files import CropFiles
