+++ mindee/v2/product/split/split_range.py
@@ -1,7 +1,7 @@
 from typing import List, Optional
 
 from mindee.v2.product.extraction.extraction_response import ExtractionResponse
-from mindee.extraction.pdf_extractor.extracted_pdf import ExtractedPdf
+from mindee.pdf.extracted_pdf import ExtractedPdf
 from mindee.input.local_input_source import LocalInputSource
 from mindee.parsing.common import StringDict
 from mindee.v2.file_operations.split import extract_single_split
