+++ tests/v1/extraction/test_image_extractor.py
@@ -3,7 +3,7 @@ import json
 import pytest
 from PIL import Image
 
-from mindee.extraction.common.image_extractor import extract_multiple_images_from_source
+from mindee.image.image_extractor import extract_multiple_images_from_source
 from mindee.input.path_input import PathInput
 from mindee.v1.product.barcode_reader import BarcodeReaderV1
 from tests.utils import V1_PRODUCT_DATA_DIR
