+++ tests/v1/extraction/test_invoice_splitter_auto_extraction.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import pytest
 
 from mindee.v1.client import Client
-from mindee.extraction.pdf_extractor.pdf_extractor import PdfExtractor
+from mindee.pdf.pdf_extractor import PdfExtractor
 from mindee.input.path_input import PathInput
 from mindee.v1.parsing.common import Document
 from mindee.v1.product.invoice.invoice_v4 import InvoiceV4
