+++ tests/v1/extraction/test_multi_receipts_extractor.py
@@ -3,9 +3,7 @@ import json
 import pytest
 from PIL import Image
 
-from mindee.extraction.multi_receipts_extractor.multi_receipts_extractor import (
-    extract_receipts,
-)
+from mindee.v1.pdf.multi_receipts_extractor import extract_receipts
 from mindee.input.path_input import PathInput
 from mindee.v1.product.multi_receipts_detector.multi_receipts_detector_v1 import (
     MultiReceiptsDetectorV1,
