+++ tests/v1/extraction/test_pdf_extractor.py
@@ -1,7 +1,7 @@
 import pytest
 
 from mindee.v1.client import Client
-from mindee.extraction.pdf_extractor.pdf_extractor import PdfExtractor
+from mindee.pdf.pdf_extractor import PdfExtractor
 from mindee.input.local_response import LocalResponse
 from mindee.input.path_input import PathInput
 from mindee.v1.product.invoice_splitter.invoice_splitter_v1 import InvoiceSplitterV1
